backend/metricMapAPI/metrics/management/future_features/predictive_maintenance.py
"""
Predictive Maintenance
======================

This module provides functions for predictive maintenance.
"""

import numpy as np
from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import train_test_split

def train_predictive_model(data, target):
    """
    Train a predictive maintenance model using RandomForest.

    Parameters:
        data (pd.DataFrame): DataFrame containing features and target.
        target (str): Target column name.

    Returns:
        model: Trained RandomForest model.
    """
    X = data.drop(columns=[target])
    y = data[target]
    
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)
    
    model = RandomForestRegressor()
    model.fit(X_train, y_train)
    
    return model

def predict_failure(model, new_data):
    """
    Predict failure using the trained model.

    Parameters:
        model (object): Trained model instance.
        new_data (pd.DataFrame): New data for making predictions.

    Returns:
        np.ndarray: Predicted failure probabilities.
    """
    predictions = model.predict(new_data)
    return predictions

# Remove unused import and trailing whitespaces.